Breaking Changes. see desc.
Removed singleton from SKSQL, now multiple databases can be opened at the same time. Some functions have been updated to receive a SKSQL instance as parameter.

SQLStatement.run now returns only one SQLResult, or JSON Array. resultTableName contains the name of the temp table for the last SELECT statement.

Tables Definitions are now cached in CTableInfoManager . Obtainable on a SKSQL instance with tableInfo. This reduces the amount of time readTableDefinition needs to be called.

Fixed Where clause: Value NOT NULL
Blocks are now marked as dirty on Insert/Update/Delete, so the SKServer can write them to disk.
Fixed TOP X: it was expecting a numeric, it now also works with a uint

The parser is being updated as to generate better error messages. more work is needed